fix: almost properly render balance
this copies over the `convert.ts` file from
stellar/soroban-example-dapp, modifying it to work with BigInt instead
of the bignumber.js library

since the balance returns the value multiplied by the number of decimal
places, i've also added `getDecimals`, which caches the value in-memory
to avoid refetching an unchanging value over and over. This inspired me
to give the same caching treatment to `getSymbol`, as well.

I'm not completely sure the BigInt vs number logic is 100% correct. The
goal is to have decimal places for small numbers, but for very large
numbers the decimal places probably aren't needed.