build: move Cargo.toml to root; rm Makefile
this is both a Rust project, for authoring contracts, and a JS project,
for creating a web app to interact with those projects. Having both a
`package.json` and a `Cargo.toml` in the project root shows that more
clearly.

Additionally, since frontend devs will probably be more familiar with
running build scripts that are defined in `package.json:scripts` rather
than a `Makefile`, the `Makefile` has been removed and the essential
items from it have been moved to `package.json`.