fix: generated package needs version for yarn
if you want to install the local package with yarn, it needs to have a
version field

unless you're using the upcoming smartdeploy tooling, versions for smart
contracts are just hashes, and even the hash isn't enough once upgrades
are allowed, so we'll need to think about that. for now, I've set the
version to `0.0.0-beta.${contract address}`, which isn't actually
compatible with semver, but it at least raises questions and will spur
more discussion around this.